+++ core/docs/changelog.md
@@ -1,4 +1,6 @@
 # TBD
+### Fixes
+* Fixed a bug where the `.kurtosis` directory wouldn't get created in the user's home directory if it didn't exist
 
 # 1.22.10
 ### Features